RTED passes on 64bit

+++ projects/RTED/CppRuntimeSystem/MemoryManager.cpp
@@ -87,7 +87,7 @@ bool MemoryType::operator< (const MemoryType & other) const
 bool MemoryType::isInitialized(int offsetFrom, int offsetTo) const
 {
     assert(offsetFrom<offsetTo);
-    assert(offsetFrom < size && offsetTo <= size);
+    assert(((unsigned int)offsetFrom) < size && ((unsigned int)offsetTo) <= size);
 
     for(int i= offsetFrom; i<offsetTo; i++)
         if(! initialized[i])
@@ -99,7 +99,7 @@ bool MemoryType::isInitialized(int offsetFrom, int offsetTo) const
 void MemoryType::initialize(int offsetFrom, int offsetTo)
 {
     assert(offsetFrom<offsetTo);
-    assert(offsetFrom < size && offsetTo <= size);
+    assert(((unsigned int)offsetFrom) < size && ((unsigned int)offsetTo) <= size);
 
     for(int i=offsetFrom; i<offsetTo; i++)
         initialized[i]=true;
